+++ backend/app.js
@@ -41,9 +41,9 @@ app.use(passport.session());
 
 // 라우터 불러오기
 require('./routes/routes.js')(app, passport);
-require('./routes/view.js')(app, passport);
-/*require('./routes/write.js')(app, passport);
-require('./routes/controller.js')(app, passport);*/
+require('./routes/view.js')(app);
+require('./routes/write.js')(app);
+//require('./routes/controller.js')(app, passport);
 
 app.listen(PORT, function() {
     console.log("Tooner Beta\n==============================================\nTooner Since 2019! Server is on at port "+PORT+"......\n==============================================");
